# openapi-spec.json
{
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/user/web3-wallet": {
      "get": {
        "description": "Get user web3 wallet detail, including total balance, ip total market cap, ip license incomes, etc...",
        "operationId": "UserController_walletDetail",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "page number",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page_size",
            "required": true,
            "in": "query",
            "description": "page size",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mint",
            "required": false,
            "in": "query",
            "description": "mint address of token to query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWalletDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user web3 wallet detail",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/v1/user/update": {
      "post": {
        "description": "update username and description",
        "operationId": "UserController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update user profile",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/v1/user/follow": {
      "post": {
        "description": "follow a user",
        "operationId": "UserController_follow",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserFollowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Follow a user",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/v1/user/unfollow": {
      "post": {
        "description": "unfollow a user",
        "operationId": "UserController_unfollow",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUnFollowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Unfollow a user",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/v1/user/sendLoginCode": {
      "post": {
        "description": "Request a login code to login with code, the code will be sent to the user's email, the code will be valid for 5 minutes",
        "operationId": "UserController_sendLoginCode",
        "parameters": [
          {
            "name": "app-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCodeReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginCodeResponseDto"
                }
              }
            }
          }
        },
        "summary": "Request a login code",
        "tags": [
          "Account"
        ]
      }
    },
...
  },
  "info": {
    "title": "3 Body Labs API Reference",
    "description": "This is the API reference for 3 Body Labs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "JWT Authorization"
      }
    },
    "schemas": {
...
      "WalletCoinSummaryDto": {
        "type": "object",
        "properties": {
          "holding_num": {
            "type": "number",
            "description": "holding number"
          },
          "formated_holding_num": {
            "type": "string",
            "description": "formated holding number"
          },
          "holding_amount": {
            "type": "number",
            "description": "holding amount"
          },
          "formated_holding_amount": {
            "type": "string",
            "description": "formated holding amount"
          },
          "price": {
            "type": "number",
            "description": "price"
          },
          "formated_price": {
            "type": "string",
            "description": "formated price"
          },
          "symbol": {
            "type": "string",
            "description": "symbol"
          },
          "name": {
            "type": "string",
            "description": "name"
          },
          "mint": {
            "type": "string",
            "description": "mint"
          },
          "cover_url": {
            "type": "string",
            "description": "cover image url"
          },
          "change5m": {
            "type": "string",
            "description": "5m change of the ip token"
          },
          "change1h": {
            "type": "string",
            "description": "1h change of the ip token"
          },
          "change24h": {
            "type": "string",
            "description": "24h change of the ip token"
          }
        },
        "required": [
          "holding_num",
          "formated_holding_num",
          "holding_amount",
          "formated_holding_amount",
          "price",
          "formated_price",
          "symbol",
          "name",
          "mint",
          "cover_url",
          "change5m",
          "change1h",
          "change24h"
        ]
      },
      "UserWalletDetailDto": {
        "type": "object",
        "properties": {
          "addr": {
            "type": "string",
            "description": "user wallet address"
          },
          "total_balance": {
            "type": "number",
            "description": "total balance"
          },
          "formated_total_balance": {
            "type": "string",
            "description": "formated total balance"
          },
          "ip_total_market_cap": {
            "type": "number",
            "description": "ip total market cap"
          },
          "formated_market_cap": {
            "type": "string",
            "description": "formated market cap"
          },
          "list": {
            "description": "list of ip tokens",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WalletCoinSummaryDto"
            }
          },
          "page": {
            "type": "number",
            "description": "page"
          },
          "page_size": {
            "type": "number",
            "description": "page size"
          },
          "ip_license_incomes": {
            "type": "number",
            "description": "ip income"
          },
          "total_balance_change_24h": {
            "type": "number",
            "description": "total balance change 24h"
          }
        },
        "required": [
          "addr",
          "total_balance",
          "formated_total_balance",
          "ip_total_market_cap",
          "formated_market_cap",
          "list",
          "page",
          "page_size",
          "ip_license_incomes",
          "total_balance_change_24h"
        ]
      },
      "UpdateProfileReqDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "description"
        ]
      },
      "UserFollowDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "description": "user of the user to follow"
          }
        },
        "required": [
          "user"
        ]
      },
      "UserUnFollowDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "description": "user of the user to follow"
          }
        },
        "required": [
          "user"
        ]
      },
      "LoginCodeReqDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "LoginCodeResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  },
...
}